refactor(settings): decompose 1262-line page into sub-components
Split src/app/settings/page.tsx into focused components:
- database-settings.tsx: backup/restore functionality
- mcp-settings.tsx: MCP access toggle with res.ok check and rollback
- backy-settings.tsx: remote backup with error toast on save failure
- two-factor-settings.tsx: TOTP setup/disable UI

page.tsx now serves as layout orchestrator (~200 lines).
