+++ src/app/settings/components/index.ts
@@ -1,0 +1,4 @@
+export { DatabaseSettings } from "./database-settings";
+export { McpSettings } from "./mcp-settings";
+export { BackySettings } from "./backy-settings";
+export { TwoFactorSettings } from "./two-factor-settings";
