+++ src/app/settings/page.tsx
@@ -1,10 +1,8 @@
 "use client";
 
-import { useState, useEffect, useCallback, useRef } from "react";
-import { useSession } from "next-auth/react";
-import { Save, Database, Bell, Shield, Terminal, Download, Upload, AlertTriangle, Loader2, Cloud, RefreshCw, History, Plug, Send, ShieldCheck, Copy, Check, Eye, EyeOff } from "lucide-react";
+import { useState, useEffect, useCallback } from "react";
+import { Save, Shield, Bell } from "lucide-react";
 import { AppShell } from "@/components/layout";
-import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
@@ -16,17 +14,13 @@ import {
   SelectValue,
 } from "@/components/ui/select";
 import { Separator } from "@/components/ui/separator";
-import { Switch } from "@/components/ui/switch";
+
 import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-} from "@/components/ui/alert-dialog";
+  DatabaseSettings,
+  McpSettings,
+  BackySettings,
+  TwoFactorSettings,
+} from "./components";
 
 interface SettingsData {
   annualIncome: string;
@@ -40,72 +34,7 @@ const DEFAULT_SETTINGS: SettingsData = {
   currency: "CNY",
 };
 
-// ── Backy types ──
-
-interface BackySettings {
-  webhookUrl: string;
-  apiKey: string;
-  hasApiKey: boolean;
-  environment: "prod" | "dev";
-}
-
-interface BackyPushResponse {
-  success: boolean;
-  error?: string;
-  request?: {
-    url: string;
-    method: string;
-    environment: string;
-    tag: string;
-    fileName: string;
-    fileSizeBytes: number;
-    backupStats: Record<string, number>;
-  };
-  response?: {
-    status: number;
-    body: unknown;
-  };
-  durationMs?: number;
-}
-
-interface BackyBackupEntry {
-  id: string;
-  tag: string;
-  environment: string;
-  file_size: number;
-  is_single_json: number;
-  created_at: string;
-}
-
-interface BackyHistoryData {
-  project_name: string;
-  environment: string | null;
-  total_backups: number;
-  recent_backups: BackyBackupEntry[];
-}
-
-// ── Backy helpers ──
-
-function formatFileSize(bytes: number): string {
-  if (bytes < 1024) return `${bytes} B`;
-  if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} KB`;
-  return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
-}
-
-function formatTimeAgo(dateStr: string): string {
-  const diff = Date.now() - new Date(dateStr).getTime();
-  const mins = Math.floor(diff / 60000);
-  if (mins < 1) return "刚刚";
-  if (mins < 60) return `${mins} 分钟前`;
-  const hours = Math.floor(mins / 60);
-  if (hours < 24) return `${hours} 小时前`;
-  const days = Math.floor(hours / 24);
-  if (days < 30) return `${days} 天前`;
-  return `${Math.floor(days / 30)} 个月前`;
-}
-
 export default function SettingsPage() {
-  const { data: sessionData, update: updateSession } = useSession();
   const [settings, setSettings] = useState<SettingsData>({
     ...DEFAULT_SETTINGS,
   });
@@ -113,42 +42,6 @@ export default function SettingsPage() {
   const [saved, setSaved] = useState(false);
   const [saving, setSaving] = useState(false);
   const [saveError, setSaveError] = useState<string | null>(null);
-  const [mcpEnabled, setMcpEnabled] = useState(false);
-  const [mcpLoading, setMcpLoading] = useState(true);
-  const [mcpPath, setMcpPath] = useState("~/workspace/personal/surety");
-
-  // Import state
-  const fileInputRef = useRef<HTMLInputElement>(null);
-  const [importDialogOpen, setImportDialogOpen] = useState(false);
-  const [importFile, setImportFile] = useState<{ name: string; data: Record<string, unknown[]> } | null>(null);
-  const [importError, setImportError] = useState<string | null>(null);
-  const [importing, setImporting] = useState(false);
-  const [importResult, setImportResult] = useState<Record<string, number> | null>(null);
-
-  // Backy state
-  const [backySettings, setBackySettings] = useState<BackySettings | null>(null);
-  const [backyUrl, setBackyUrl] = useState("");
-  const [backyKey, setBackyKey] = useState("");
-  const [backySaving, setBackySaving] = useState(false);
-  const [backyTesting, setBackyTesting] = useState(false);
-  const [backyTestResult, setBackyTestResult] = useState<{ success: boolean; status: number; error?: string } | null>(null);
-  const [backyPushing, setBackyPushing] = useState(false);
-  const [backyPushResult, setBackyPushResult] = useState<BackyPushResponse | null>(null);
-  const [backyHistory, setBackyHistory] = useState<BackyHistoryData | null>(null);
-  const [backyHistoryLoading, setBackyHistoryLoading] = useState(false);
-
-  // 2FA state
-  const [tfaEnabled, setTfaEnabled] = useState(false);
-  const [tfaLoading, setTfaLoading] = useState(true);
-  const [tfaSetupData, setTfaSetupData] = useState<{ qrDataURL: string; secret: string } | null>(null);
-  const [tfaSetupCode, setTfaSetupCode] = useState("");
-  const [tfaDisableCode, setTfaDisableCode] = useState("");
-  const [tfaRecoveryCode, setTfaRecoveryCode] = useState<string | null>(null);
-  const [tfaError, setTfaError] = useState<string | null>(null);
-  const [tfaProcessing, setTfaProcessing] = useState(false);
-  const [tfaShowSecret, setTfaShowSecret] = useState(false);
-  const [tfaCopied, setTfaCopied] = useState(false);
-  const [tfaShowDisable, setTfaShowDisable] = useState(false);
 
   const loadSettings = useCallback(async () => {
     try {
@@ -179,346 +72,9 @@ export default function SettingsPage() {
     }
   }, []);
 
-  // Load MCP setting from backend
-  const loadMcpSetting = useCallback(async () => {
-    try {
-      const res = await fetch("/api/settings/mcp.enabled");
-      if (res.ok) {
-        const data = await res.json();
-        setMcpEnabled(data.value === "true");
-      }
-    } catch {
-      // Setting doesn't exist yet, default to false
-    } finally {
-      setMcpLoading(false);
-    }
-  }, []);
-
-  // Load Backy settings from backend
-  const loadBackySettings = useCallback(async () => {
-    try {
-      const res = await fetch("/api/settings/backy");
-      if (res.ok) {
-        const data: BackySettings = await res.json();
-        setBackySettings(data);
-        setBackyUrl(data.webhookUrl);
-        // Don't overwrite key input with masked value
-      }
-    } catch {
-      // ignore
-    }
-  }, []);
-
-  // Load Backy history
-  const loadBackyHistory = useCallback(async () => {
-    setBackyHistoryLoading(true);
-    try {
-      const res = await fetch("/api/settings/backy/history");
-      if (res.ok) {
-        const data: BackyHistoryData = await res.json();
-        setBackyHistory(data);
-      }
-    } catch {
-      // ignore
-    } finally {
-      setBackyHistoryLoading(false);
-    }
-  }, []);
-
-  // Load 2FA status
-  const loadTfaStatus = useCallback(async () => {
-    try {
-      const res = await fetch("/api/settings/2fa/status");
-      if (res.ok) {
-        const data = await res.json();
-        setTfaEnabled(data.enabled);
-      }
-    } catch {
-      // ignore
-    } finally {
-      setTfaLoading(false);
-    }
-  }, []);
-
-  // 2FA: start setup
-  const handleTfaSetup = async () => {
-    setTfaProcessing(true);
-    setTfaError(null);
-    try {
-      const res = await fetch("/api/settings/2fa/setup", { method: "POST" });
-      const data = await res.json();
-      if (!res.ok) {
-        setTfaError(data.error || "Setup failed");
-        return;
-      }
-      setTfaSetupData(data);
-      setTfaSetupCode("");
-    } catch {
-      setTfaError("Network error");
-    } finally {
-      setTfaProcessing(false);
-    }
-  };
-
-  // 2FA: verify setup code
-  const handleTfaVerifySetup = async () => {
-    if (!tfaSetupCode || !/^\d{6}$/.test(tfaSetupCode)) {
-      setTfaError("Please enter a 6-digit code");
-      return;
-    }
-    setTfaProcessing(true);
-    setTfaError(null);
-    try {
-      const res = await fetch("/api/settings/2fa/verify-setup", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ token: tfaSetupCode }),
-      });
-      const data = await res.json();
-      if (!res.ok) {
-        setTfaError(data.error || "Verification failed");
-        return;
-      }
-      setTfaEnabled(true);
-      setTfaRecoveryCode(data.recoveryCode);
-      setTfaSetupData(null);
-      setTfaSetupCode("");
-
-      // Promote JWT: setup proves authenticator ownership, exempt current session
-      if (data.twoFactorNonce && data.twoFactorSig) {
-        await updateSession({
-          twoFactorNonce: data.twoFactorNonce,
-          twoFactorSig: data.twoFactorSig,
-        });
-      }
-    } catch {
-      setTfaError("Network error");
-    } finally {
-      setTfaProcessing(false);
-    }
-  };
-
-  // 2FA: disable
-  const handleTfaDisable = async () => {
-    if (!tfaDisableCode || !/^\d{6}$/.test(tfaDisableCode)) {
-      setTfaError("Please enter a 6-digit code");
-      return;
-    }
-    setTfaProcessing(true);
-    setTfaError(null);
-    try {
-      const res = await fetch("/api/settings/2fa/disable", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ token: tfaDisableCode }),
-      });
-      const data = await res.json();
-      if (!res.ok) {
-        setTfaError(data.error || "Failed to disable 2FA");
-        return;
-      }
-      if (data.success) {
-        setTfaEnabled(false);
-        setTfaShowDisable(false);
-        setTfaDisableCode("");
-        setTfaRecoveryCode(null);
-      }
-    } catch {
-      setTfaError("Network error");
-    } finally {
-      setTfaProcessing(false);
-    }
-  };
-
-  // 2FA: force disable (recovery session — no authenticator available)
-  const handleTfaForceDisable = async () => {
-    setTfaProcessing(true);
-    setTfaError(null);
-    try {
-      const res = await fetch("/api/settings/2fa/disable", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ force: true }),
-      });
-      const data = await res.json();
-      if (!res.ok) {
-        setTfaError(data.error || "Failed to disable 2FA");
-        return;
-      }
-      if (data.success) {
-        setTfaEnabled(false);
-        setTfaShowDisable(false);
-        setTfaDisableCode("");
-        setTfaRecoveryCode(null);
-        // Revoke the one-time force-disable privilege from JWT
-        if (data.clearRecoverySession) {
-          await updateSession({ clearRecoverySession: true });
-        }
-      }
-    } catch {
-      setTfaError("Network error");
-    } finally {
-      setTfaProcessing(false);
-    }
-  };
-
-  // 2FA: copy recovery code
-  const handleCopyRecoveryCode = async () => {
-    if (!tfaRecoveryCode) return;
-    await navigator.clipboard.writeText(tfaRecoveryCode);
-    setTfaCopied(true);
-    setTimeout(() => setTfaCopied(false), 2000);
-  };
-
   useEffect(() => {
     void loadSettings();
-    void loadMcpSetting();
-    void loadTfaStatus();
-    void loadBackySettings().then(() => {
-      // Auto-load history after settings are loaded (if configured)
-    });
-  }, [loadSettings, loadMcpSetting, loadTfaStatus, loadBackySettings]);
-
-  // Auto-load history when backy settings are loaded and configured
-  useEffect(() => {
-    if (backySettings?.hasApiKey && backySettings.webhookUrl) {
-      void loadBackyHistory();
-    }
-  }, [backySettings, loadBackyHistory]);
-
-  // Toggle MCP setting via API
-  const handleMcpToggle = async (enabled: boolean) => {
-    setMcpEnabled(enabled);
-    try {
-      await fetch("/api/settings", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ key: "mcp.enabled", value: String(enabled) }),
-      });
-    } catch {
-      // Revert on failure
-      setMcpEnabled(!enabled);
-    }
-  };
-
-  // Table display names for import preview
-  const TABLE_LABELS: Record<string, string> = {
-    members: "成员",
-    insurers: "保险公司",
-    assets: "资产",
-    policies: "保单",
-    beneficiaries: "受益人",
-    payments: "缴费记录",
-    cashValues: "现金价值",
-    settings: "设置",
-  };
-
-  // Handle file selection for import
-  const handleFileSelect = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setImportError(null);
-    setImportResult(null);
-    const file = e.target.files?.[0];
-    if (!file) return;
-
-    const reader = new FileReader();
-    reader.onload = (ev) => {
-      try {
-        const json = JSON.parse(ev.target?.result as string);
-        if (json.version !== 1 || !json.data) {
-          setImportError("无效的备份文件：缺少 version 或 data 字段");
-          return;
-        }
-        setImportFile({ name: file.name, data: json.data });
-        setImportDialogOpen(true);
-      } catch {
-        setImportError("无法解析 JSON 文件，请检查文件格式");
-      }
-    };
-    reader.readAsText(file);
-    // Reset input so the same file can be re-selected
-    e.target.value = "";
-  };
-
-  // Execute the actual import
-  const handleImportConfirm = async () => {
-    if (!importFile) return;
-    setImporting(true);
-    setImportError(null);
-    try {
-      const res = await fetch("/api/backup", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ version: 1, data: importFile.data }),
-      });
-      const body = await res.json();
-      if (!res.ok) {
-        setImportError(body.error || "导入失败");
-      } else {
-        setImportResult(body.restored);
-      }
-    } catch {
-      setImportError("网络错误，请重试");
-    } finally {
-      setImporting(false);
-      setImportDialogOpen(false);
-      setImportFile(null);
-    }
-  };
-
-  // Backy: save settings
-  const handleBackySave = async () => {
-    setBackySaving(true);
-    setBackyTestResult(null);
-    try {
-      const res = await fetch("/api/settings/backy", {
-        method: "PUT",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ webhookUrl: backyUrl, apiKey: backyKey }),
-      });
-      if (res.ok) {
-        const data: BackySettings = await res.json();
-        setBackySettings(data);
-        setBackyKey(""); // Clear key input after save
-      }
-    } catch {
-      // ignore
-    } finally {
-      setBackySaving(false);
-    }
-  };
-
-  // Backy: test connection
-  const handleBackyTest = async () => {
-    setBackyTesting(true);
-    setBackyTestResult(null);
-    try {
-      const res = await fetch("/api/settings/backy/test", { method: "POST" });
-      const data = await res.json();
-      setBackyTestResult(data);
-    } catch {
-      setBackyTestResult({ success: false, status: 0, error: "网络错误" });
-    } finally {
-      setBackyTesting(false);
-    }
-  };
-
-  // Backy: push backup
-  const handleBackyPush = async () => {
-    setBackyPushing(true);
-    setBackyPushResult(null);
-    try {
-      const res = await fetch("/api/settings/backy/push", { method: "POST" });
-      const data: BackyPushResponse = await res.json();
-      setBackyPushResult(data);
-      if (data.success) {
-        void loadBackyHistory();
-      }
-    } catch {
-      setBackyPushResult({ success: false, error: "网络错误" });
-    } finally {
-      setBackyPushing(false);
-    }
-  };
+  }, [loadSettings]);
 
   const handleChange = (field: keyof SettingsData, value: string) => {
     setSettings((prev) => ({ ...prev, [field]: value }));
@@ -562,8 +118,6 @@ export default function SettingsPage() {
     }
   };
 
-  const backyConfigured = backySettings?.hasApiKey && backySettings.webhookUrl;
-
   return (
     <AppShell breadcrumbs={[{ label: "设置" }]}>
       <div className="space-y-6">
@@ -575,6 +129,7 @@ export default function SettingsPage() {
         </div>
 
         <div className="grid gap-6 lg:grid-cols-2">
+          {/* Family Finance Settings */}
           <div className="rounded-card bg-secondary p-6">
             <div className="flex items-center gap-3 mb-4">
               <div className="flex h-10 w-10 items-center justify-center rounded-lg bg-primary/10">
@@ -620,6 +175,7 @@ export default function SettingsPage() {
             </div>
           </div>
 
+          {/* Reminder Settings */}
           <div className="rounded-card bg-secondary p-6">
             <div className="flex items-center gap-3 mb-4">
               <div className="flex h-10 w-10 items-center justify-center rounded-lg bg-orange-500/10">
@@ -655,594 +211,17 @@ export default function SettingsPage() {
             </div>
           </div>
 
-          <div className="rounded-card bg-secondary p-6">
-            <div className="flex items-center gap-3 mb-4">
-              <div className="flex h-10 w-10 items-center justify-center rounded-lg bg-blue-500/10">
-                <Database className="h-5 w-5 text-blue-500" strokeWidth={1.5} />
-              </div>
-              <div>
-                <h2 className="font-semibold">数据管理</h2>
-                <p className="text-sm text-muted-foreground">备份与恢复</p>
-              </div>
-            </div>
-            <Separator className="mb-4" />
-            <div className="space-y-4">
-              <div className="space-y-2">
-                <p className="text-sm text-muted-foreground">
-                  导出所有保单、成员、资产等数据为 JSON 文件，可用于备份或迁移。
-                </p>
-                <Button
-                  variant="outline"
-                  size="sm"
-                  onClick={() => {
-                    window.location.href = "/api/backup";
-                  }}
-                >
-                  <Download className="mr-2 h-4 w-4" />
-                  导出 JSON 备份
-                </Button>
-              </div>
+          {/* Database Settings (backup/restore) */}
+          <DatabaseSettings />
 
-              <Separator />
-
-              <div className="space-y-2">
-                <p className="text-sm text-muted-foreground">
-                  从 JSON 备份文件恢复数据。此操作将覆盖当前所有数据。
-                </p>
-                <input
-                  ref={fileInputRef}
-                  type="file"
-                  accept=".json"
-                  className="hidden"
-                  onChange={handleFileSelect}
-                />
-                <Button
-                  variant="outline"
-                  size="sm"
-                  onClick={() => fileInputRef.current?.click()}
-                >
-                  <Upload className="mr-2 h-4 w-4" />
-                  导入 JSON 备份
-                </Button>
-
-                {importError && (
-                  <div className="flex items-start gap-2 rounded-widget border border-red-200 bg-red-50 p-3 dark:border-red-900 dark:bg-red-950">
-                    <AlertTriangle className="h-4 w-4 text-red-600 dark:text-red-400 shrink-0 mt-0.5" />
-                    <p className="text-sm text-red-700 dark:text-red-300">{importError}</p>
-                  </div>
-                )}
-
-                {importResult && (
-                  <div className="rounded-widget border border-green-200 bg-green-50 p-3 dark:border-green-900 dark:bg-green-950">
-                    <p className="text-sm font-medium text-green-800 dark:text-green-200 mb-1">
-                      导入成功
-                    </p>
-                    <div className="text-xs text-green-700 dark:text-green-300 grid grid-cols-3 gap-x-4 gap-y-1">
-                      {Object.entries(importResult).map(([key, count]) => (
-                        <span key={key}>
-                          {TABLE_LABELS[key] || key}: {count}
-                        </span>
-                      ))}
-                    </div>
-                  </div>
-                )}
-              </div>
-            </div>
-
-            <AlertDialog open={importDialogOpen} onOpenChange={setImportDialogOpen}>
-              <AlertDialogContent>
-                <AlertDialogHeader>
-                  <AlertDialogTitle>确认导入数据？</AlertDialogTitle>
-                  <AlertDialogDescription>
-                    此操作将<span className="font-semibold text-foreground">覆盖当前所有数据</span>，不可撤销。
-                    请确认已备份重要数据。
-                  </AlertDialogDescription>
-                </AlertDialogHeader>
-
-                {importFile && (
-                  <div className="rounded-widget border bg-muted/50 p-3 text-sm">
-                    <p className="font-medium mb-2">文件: {importFile.name}</p>
-                    <div className="grid grid-cols-3 gap-x-4 gap-y-1 text-muted-foreground">
-                      {Object.entries(importFile.data).map(([key, rows]) => (
-                        <span key={key}>
-                          {TABLE_LABELS[key] || key}: {Array.isArray(rows) ? rows.length : 0}
-                        </span>
-                      ))}
-                    </div>
-                  </div>
-                )}
-
-                <AlertDialogFooter>
-                  <AlertDialogCancel disabled={importing}>取消</AlertDialogCancel>
-                  <AlertDialogAction
-                    variant="destructive"
-                    onClick={handleImportConfirm}
-                    disabled={importing}
-                  >
-                    {importing && <Loader2 className="mr-2 h-4 w-4 animate-spin" />}
-                    确认覆盖导入
-                  </AlertDialogAction>
-                </AlertDialogFooter>
-              </AlertDialogContent>
-            </AlertDialog>
-          </div>
-
-          {/* MCP Access */}
-          <div className="rounded-card bg-secondary p-6">
-            <div className="flex items-center justify-between">
-              <div className="flex items-center gap-3">
-                <div className="flex h-10 w-10 items-center justify-center rounded-lg bg-emerald-500/10">
-                  <Terminal className="h-5 w-5 text-emerald-500" strokeWidth={1.5} />
-                </div>
-                <div>
-                  <h2 className="font-semibold">MCP 访问</h2>
-                  <p className="text-sm text-muted-foreground">
-                    允许外部 AI 助手（Claude Code、Cursor 等）查询保单数据
-                  </p>
-                </div>
-              </div>
-              <Switch
-                checked={mcpEnabled}
-                onCheckedChange={handleMcpToggle}
-                disabled={mcpLoading}
-                aria-label="Toggle MCP access"
-              />
-            </div>
-            {mcpEnabled && (
-              <>
-                <Separator className="my-4" />
-                <div className="rounded-widget bg-muted/50 p-4">
-                  <p className="text-sm font-medium mb-2">配置方式</p>
-                  <p className="text-xs text-muted-foreground mb-3">
-                    在 AI 助手的 MCP 配置中添加以下内容：
-                  </p>
-                  <div className="mb-3">
-                    <Label htmlFor="mcp-path" className="text-xs text-muted-foreground mb-1 block">
-                      项目路径
-                    </Label>
-                    <Input
-                      id="mcp-path"
-                      value={mcpPath}
-                      onChange={(e) => setMcpPath(e.target.value)}
-                      className="text-xs font-mono h-8"
-                      placeholder="~/workspace/personal/surety"
-                    />
-                  </div>
-                  <pre className="text-xs bg-background rounded-widget p-3 overflow-x-auto">
-{`{
-  "mcpServers": {
-    "surety": {
-      "command": "bun",
-      "args": ["run", "${mcpPath}/mcp/index.ts"],
-      "cwd": "${mcpPath}"
-    }
-  }
-}`}
-                  </pre>
-                </div>
-              </>
-            )}
-          </div>
+          {/* MCP Access Settings */}
+          <McpSettings />
 
           {/* Two-Factor Authentication */}
-          <div className="rounded-card bg-secondary p-6">
-            <div className="flex items-center gap-3 mb-4">
-              <div className="flex h-10 w-10 items-center justify-center rounded-lg bg-amber-500/10">
-                <ShieldCheck className="h-5 w-5 text-amber-500" strokeWidth={1.5} />
-              </div>
-              <div>
-                <h2 className="font-semibold">双因素认证</h2>
-                <p className="text-sm text-muted-foreground">
-                  {tfaEnabled ? "已启用 TOTP 双因素认证" : "使用 TOTP 增强登录安全性"}
-                </p>
-              </div>
-              {tfaEnabled && (
-                <Badge variant="success" className="ml-auto">已启用</Badge>
-              )}
-            </div>
-            <Separator className="mb-4" />
-
-            {tfaLoading ? (
-              <div className="flex items-center gap-2 text-sm text-muted-foreground">
-                <Loader2 className="h-4 w-4 animate-spin" />
-                加载中...
-              </div>
-            ) : !tfaEnabled ? (
-              /* --- NOT ENABLED: setup flow --- */
-              <div className="space-y-4">
-                {!tfaSetupData ? (
-                  /* Step 1: start setup */
-                  <>
-                    <p className="text-sm text-muted-foreground">
-                      启用后，每次登录需输入 Google Authenticator 等应用生成的 6 位验证码。
-                    </p>
-                    <Button
-                      variant="outline"
-                      size="sm"
-                      onClick={() => void handleTfaSetup()}
-                      disabled={tfaProcessing}
-                    >
-                      {tfaProcessing && <Loader2 className="mr-2 h-4 w-4 animate-spin" />}
-                      <ShieldCheck className="mr-2 h-4 w-4" />
-                      开始设置
-                    </Button>
-                  </>
-                ) : (
-                  /* Step 2: scan QR and verify */
-                  <>
-                    <p className="text-sm text-muted-foreground">
-                      使用 Google Authenticator 或其他 TOTP 应用扫描二维码，然后输入 6 位验证码确认。
-                    </p>
-                    {/* QR code */}
-                    <div>
-                      {/* eslint-disable-next-line @next/next/no-img-element */}
-                      <img
-                        src={tfaSetupData.qrDataURL}
-                        alt="TOTP QR Code"
-                        className="h-48 w-48 rounded-lg border bg-white p-2"
-                      />
-                    </div>
-                    {/* Manual secret */}
-                    <div className="space-y-1">
-                      <div className="flex items-center gap-2">
-                        <Label className="text-xs text-muted-foreground">Manual key</Label>
-                        <button
-                          onClick={() => setTfaShowSecret(!tfaShowSecret)}
-                          className="text-xs text-muted-foreground hover:text-foreground transition-colors cursor-pointer"
-                        >
-                          {tfaShowSecret ? <EyeOff className="h-3 w-3" /> : <Eye className="h-3 w-3" />}
-                        </button>
-                      </div>
-                      {tfaShowSecret && (
-                        <code className="block text-xs font-mono bg-muted rounded-widget px-2 py-1 break-all select-all">
-                          {tfaSetupData.secret}
-                        </code>
-                      )}
-                    </div>
-                    {/* Verify code input */}
-                    <div className="flex items-center gap-2">
-                      <Input
-                        className="h-9 w-40 font-mono text-center tracking-widest"
-                        placeholder="000000"
-                        maxLength={6}
-                        value={tfaSetupCode}
-                        onChange={(e) => setTfaSetupCode(e.target.value.replace(/\D/g, "").slice(0, 6))}
-                      />
-                      <Button
-                        size="sm"
-                        onClick={() => void handleTfaVerifySetup()}
-                        disabled={tfaProcessing || tfaSetupCode.length !== 6}
-                      >
-                        {tfaProcessing && <Loader2 className="mr-2 h-4 w-4 animate-spin" />}
-                        确认启用
-                      </Button>
-                    </div>
-                  </>
-                )}
-
-                {/* Recovery code display (shown right after enabling) */}
-                {tfaRecoveryCode && (
-                  <div className="rounded-widget border border-amber-200 bg-amber-50 p-4 dark:border-amber-900 dark:bg-amber-950 space-y-2">
-                    <p className="text-sm font-medium text-amber-800 dark:text-amber-200">
-                      Recovery Code (save this — shown only once)
-                    </p>
-                    <div className="flex items-center gap-2">
-                      <code className="text-sm font-mono bg-background rounded-widget px-3 py-1.5 select-all">
-                        {tfaRecoveryCode}
-                      </code>
-                      <button
-                        onClick={() => void handleCopyRecoveryCode()}
-                        className="text-muted-foreground hover:text-foreground transition-colors cursor-pointer"
-                        aria-label="Copy recovery code"
-                      >
-                        {tfaCopied ? <Check className="h-4 w-4 text-green-600" /> : <Copy className="h-4 w-4" />}
-                      </button>
-                    </div>
-                    <p className="text-xs text-amber-700 dark:text-amber-300">
-                      如果您丢失了认证设备，可以使用此备用码登录（仅可使用一次）。
-                    </p>
-                  </div>
-                )}
-
-                {tfaError && (
-                  <div className="flex items-start gap-2 rounded-widget border border-red-200 bg-red-50 p-3 dark:border-red-900 dark:bg-red-950">
-                    <AlertTriangle className="h-4 w-4 text-red-600 dark:text-red-400 shrink-0 mt-0.5" />
-                    <p className="text-sm text-red-700 dark:text-red-300">{tfaError}</p>
-                  </div>
-                )}
-              </div>
-            ) : (
-              /* --- ENABLED: status & disable --- */
-              <div className="space-y-4">
-                <div className="text-sm text-muted-foreground space-y-1">
-                  <p>TOTP 双因素认证已启用。每次登录需要验证码。</p>
-                  {sessionData?.user?.recoverySession && (
-                    <p className="text-amber-600 dark:text-amber-400">
-                      当前会话通过 Recovery code 登录。建议禁用后重新启用以获取新的 recovery code。
-                    </p>
-                  )}
-                </div>
-
-                {/* Recovery code display (if just enabled) */}
-                {tfaRecoveryCode && (
-                  <div className="rounded-widget border border-amber-200 bg-amber-50 p-4 dark:border-amber-900 dark:bg-amber-950 space-y-2">
-                    <p className="text-sm font-medium text-amber-800 dark:text-amber-200">
-                      Recovery Code (save this — shown only once)
-                    </p>
-                    <div className="flex items-center gap-2">
-                      <code className="text-sm font-mono bg-background rounded-widget px-3 py-1.5 select-all">
-                        {tfaRecoveryCode}
-                      </code>
-                      <button
-                        onClick={() => void handleCopyRecoveryCode()}
-                        className="text-muted-foreground hover:text-foreground transition-colors cursor-pointer"
-                        aria-label="Copy recovery code"
-                      >
-                        {tfaCopied ? <Check className="h-4 w-4 text-green-600" /> : <Copy className="h-4 w-4" />}
-                      </button>
-                    </div>
-                    <p className="text-xs text-amber-700 dark:text-amber-300">
-                      如果您丢失了认证设备，可以使用此备用码登录（仅可使用一次）。
-                    </p>
-                  </div>
-                )}
-
-                {!tfaShowDisable ? (
-                  <Button
-                    variant="outline"
-                    size="sm"
-                    onClick={() => { setTfaShowDisable(true); setTfaError(null); }}
-                  >
-                    禁用双因素认证
-                  </Button>
-                ) : sessionData?.user?.recoverySession ? (
-                  /* Recovery session — authenticator lost, allow force disable */
-                  <div className="space-y-3 rounded-widget border border-destructive/20 bg-destructive/5 p-4">
-                    <p className="text-sm font-medium text-destructive">
-                      当前会话通过 Recovery code 登录，您可以直接禁用双因素认证。禁用后建议重新启用以获取新的 recovery code。
-                    </p>
-                    <div className="flex items-center gap-2">
-                      <Button
-                        variant="destructive"
-                        size="sm"
-                        onClick={() => void handleTfaForceDisable()}
-                        disabled={tfaProcessing}
-                      >
-                        {tfaProcessing && <Loader2 className="mr-2 h-4 w-4 animate-spin" />}
-                        确认禁用
-                      </Button>
-                      <Button
-                        variant="ghost"
-                        size="sm"
-                        onClick={() => { setTfaShowDisable(false); setTfaError(null); }}
-                      >
-                        取消
-                      </Button>
-                    </div>
-                  </div>
-                ) : (
-                  <div className="space-y-3 rounded-widget border border-destructive/20 bg-destructive/5 p-4">
-                    <p className="text-sm font-medium text-destructive">
-                      输入当前验证码以禁用双因素认证
-                    </p>
-                    <div className="flex items-center gap-2">
-                      <Input
-                        className="h-9 w-40 font-mono text-center tracking-widest"
-                        placeholder="000000"
-                        maxLength={6}
-                        value={tfaDisableCode}
-                        onChange={(e) => setTfaDisableCode(e.target.value.replace(/\D/g, "").slice(0, 6))}
-                      />
-                      <Button
-                        variant="destructive"
-                        size="sm"
-                        onClick={() => void handleTfaDisable()}
-                        disabled={tfaProcessing || tfaDisableCode.length !== 6}
-                      >
-                        {tfaProcessing && <Loader2 className="mr-2 h-4 w-4 animate-spin" />}
-                        确认禁用
-                      </Button>
-                      <Button
-                        variant="ghost"
-                        size="sm"
-                        onClick={() => { setTfaShowDisable(false); setTfaDisableCode(""); setTfaError(null); }}
-                      >
-                        取消
-                      </Button>
-                    </div>
-                  </div>
-                )}
-
-                {tfaError && (
-                  <div className="flex items-start gap-2 rounded-widget border border-red-200 bg-red-50 p-3 dark:border-red-900 dark:bg-red-950">
-                    <AlertTriangle className="h-4 w-4 text-red-600 dark:text-red-400 shrink-0 mt-0.5" />
-                    <p className="text-sm text-red-700 dark:text-red-300">{tfaError}</p>
-                  </div>
-                )}
-              </div>
-            )}
-          </div>
+          <TwoFactorSettings />
 
           {/* Backy Remote Backup */}
-          <div className="rounded-card bg-secondary p-6">
-            <div className="flex items-center gap-3 mb-4">
-              <div className="flex h-10 w-10 items-center justify-center rounded-lg bg-violet-500/10">
-                <Cloud className="h-5 w-5 text-violet-500" strokeWidth={1.5} />
-              </div>
-              <div className="flex items-center gap-2">
-                <div>
-                  <h2 className="font-semibold">远程备份</h2>
-                  <p className="text-sm text-muted-foreground">推送数据到 Backy 服务</p>
-                </div>
-                <Badge variant={backySettings?.environment === "prod" ? "success" : "warning"} className="ml-1">
-                  {backySettings?.environment ?? "dev"}
-                </Badge>
-              </div>
-            </div>
-            <Separator className="mb-4" />
-
-            {/* Config form */}
-            <div className="space-y-4">
-              <div className="space-y-1">
-                <Label className="text-sm">Webhook URL</Label>
-                <Input
-                  className="h-9"
-                  placeholder="https://backy.example.com/api/webhook/..."
-                  value={backyUrl}
-                  onChange={(e) => setBackyUrl(e.target.value)}
-                />
-              </div>
-              <div className="space-y-1">
-                <Label className="text-sm">API Key</Label>
-                <Input
-                  className="h-9"
-                  type="password"
-                  placeholder={backySettings?.hasApiKey ? "已配置（输入新值可覆盖）" : "输入 API Key"}
-                  value={backyKey}
-                  onChange={(e) => setBackyKey(e.target.value)}
-                />
-              </div>
-
-              {/* Action buttons */}
-              <div className="flex items-center gap-2">
-                <Button
-                  variant="outline"
-                  size="sm"
-                  onClick={handleBackySave}
-                  disabled={backySaving || !backyUrl}
-                >
-                  {backySaving && <Loader2 className="mr-2 h-4 w-4 animate-spin" />}
-                  <Save className="mr-2 h-4 w-4" />
-                  保存
-                </Button>
-                <Button
-                  variant="outline"
-                  size="sm"
-                  onClick={handleBackyTest}
-                  disabled={backyTesting || !backyConfigured}
-                >
-                  {backyTesting && <Loader2 className="mr-2 h-4 w-4 animate-spin" />}
-                  <Plug className="mr-2 h-4 w-4" />
-                  测试连接
-                </Button>
-                <Button
-                  variant="outline"
-                  size="sm"
-                  onClick={handleBackyPush}
-                  disabled={backyPushing || !backyConfigured}
-                >
-                  {backyPushing && <Loader2 className="mr-2 h-4 w-4 animate-spin" />}
-                  <Send className="mr-2 h-4 w-4" />
-                  推送备份
-                </Button>
-              </div>
-
-              {/* Test result */}
-              {backyTestResult && (
-                <div className={`flex items-start gap-2 rounded-widget border p-3 ${
-                  backyTestResult.success
-                    ? "border-green-200 bg-green-50 dark:border-green-900 dark:bg-green-950"
-                    : "border-red-200 bg-red-50 dark:border-red-900 dark:bg-red-950"
-                }`}>
-                  {backyTestResult.success ? (
-                    <p className="text-sm text-green-700 dark:text-green-300">
-                      连接成功 (HTTP {backyTestResult.status})
-                    </p>
-                  ) : (
-                    <>
-                      <AlertTriangle className="h-4 w-4 text-red-600 dark:text-red-400 shrink-0 mt-0.5" />
-                      <p className="text-sm text-red-700 dark:text-red-300">
-                        连接失败{backyTestResult.status ? ` (HTTP ${backyTestResult.status})` : ""}{backyTestResult.error ? `: ${backyTestResult.error}` : ""}
-                      </p>
-                    </>
-                  )}
-                </div>
-              )}
-
-              {/* Push result */}
-              {backyPushResult && (
-                <div className={`rounded-widget border p-3 ${
-                  backyPushResult.success
-                    ? "border-green-200 bg-green-50 dark:border-green-900 dark:bg-green-950"
-                    : "border-red-200 bg-red-50 dark:border-red-900 dark:bg-red-950"
-                }`}>
-                  {backyPushResult.success ? (
-                    <div className="space-y-2">
-                      <p className="text-sm font-medium text-green-800 dark:text-green-200">
-                        推送成功 ({backyPushResult.durationMs}ms)
-                      </p>
-                      {backyPushResult.request && (
-                        <div className="text-xs text-green-700 dark:text-green-300 space-y-1">
-                          <p>Tag: {backyPushResult.request.tag}</p>
-                          <p>文件: {backyPushResult.request.fileName} ({formatFileSize(backyPushResult.request.fileSizeBytes)})</p>
-                          <div className="grid grid-cols-3 gap-x-4 gap-y-0.5">
-                            {Object.entries(backyPushResult.request.backupStats).map(([key, count]) => (
-                              <span key={key}>{TABLE_LABELS[key] || key}: {count}</span>
-                            ))}
-                          </div>
-                        </div>
-                      )}
-                    </div>
-                  ) : (
-                    <div className="flex items-start gap-2">
-                      <AlertTriangle className="h-4 w-4 text-red-600 dark:text-red-400 shrink-0 mt-0.5" />
-                      <div className="text-sm text-red-700 dark:text-red-300">
-                        <p>推送失败: {backyPushResult.error}</p>
-                        {backyPushResult.response && (
-                          <p className="text-xs mt-1">HTTP {backyPushResult.response.status}: {JSON.stringify(backyPushResult.response.body)}</p>
-                        )}
-                      </div>
-                    </div>
-                  )}
-                </div>
-              )}
-            </div>
-
-            {/* Remote history panel */}
-            {backyConfigured && (
-              <>
-                <Separator className="my-4" />
-                <div className="space-y-3">
-                  <div className="flex items-center justify-between">
-                    <div className="flex items-center gap-2">
-                      <History className="h-4 w-4 text-muted-foreground" />
-                      <span className="text-sm font-medium">远程备份记录</span>
-                      {backyHistory && (
-                        <Badge variant="secondary">{backyHistory.total_backups} 份</Badge>
-                      )}
-                    </div>
-                    <Button
-                      variant="ghost"
-                      size="sm"
-                      onClick={() => void loadBackyHistory()}
-                      disabled={backyHistoryLoading}
-                    >
-                      <RefreshCw className={`h-4 w-4 ${backyHistoryLoading ? "animate-spin" : ""}`} />
-                    </Button>
-                  </div>
-
-                  {backyHistory && backyHistory.recent_backups.length > 0 ? (
-                    <div className="grid gap-2 sm:grid-cols-2">
-                      {backyHistory.recent_backups.map((entry) => (
-                        <div key={entry.id} className="rounded-widget border bg-muted/50 p-3 text-xs space-y-1">
-                          <div className="flex items-center justify-between">
-                            <Badge variant={entry.environment === "prod" ? "success" : "warning"}>
-                              {entry.environment}
-                            </Badge>
-                            <span className="text-muted-foreground">{formatFileSize(entry.file_size)}</span>
-                          </div>
-                          <p className="text-muted-foreground truncate" title={entry.tag}>{entry.tag}</p>
-                          <p className="text-muted-foreground">{formatTimeAgo(entry.created_at)}</p>
-                        </div>
-                      ))}
-                    </div>
-                  ) : backyHistory && backyHistory.recent_backups.length === 0 ? (
-                    <p className="text-xs text-muted-foreground">暂无备份记录</p>
-                  ) : null}
-                </div>
-              </>
-            )}
-          </div>
+          <BackySettings />
         </div>
 
         <div className="flex flex-col items-end gap-2">
